+++ src/multiselect_patch.js
