+++ src/multiselect_shortcut.js
